# backend/docs/swagger.yaml
components:
  schemas:
...
    handler.createTicketTypeBody:
      properties:
        capacity:
          type: integer
        description:
          type: string
        name:
          type: string
        price_cents:
          type: integer
      type: object
...
    handler.updateEventBody:
      properties:
        cover_image_key:
          type: string
        description:
          type: string
        ends_at:
          type: string
        name:
          type: string
        slug:
          type: string
        starts_at:
          type: string
        timezone:
          type: string
        venue_address:
          type: string
        venue_name:
          type: string
      type: object
    handler.updateTicketTypeBody:
      properties:
        capacity:
          type: integer
        description:
          type: string
        name:
          type: string
        price_cents:
          type: integer
        sort_order:
          type: integer
      type: object
    internal_catalog_openapi.MessageData:
      properties:
        message:
          type: string
      type: object
...
    openapi.EnvelopeEventDetail:
      properties:
        data:
          $ref: '#/components/schemas/service.EventDetail'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    openapi.EnvelopeMessage:
      properties:
        data:
          $ref: '#/components/schemas/internal_catalog_openapi.MessageData'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    openapi.EnvelopeTicketTypeDetail:
      properties:
        data:
          $ref: '#/components/schemas/service.TicketTypeDetail'
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
    openapi.EnvelopeTicketTypeList:
      properties:
        data:
          items:
            $ref: '#/components/schemas/service.TicketTypeDetail'
          type: array
          uniqueItems: false
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    platform.APIError:
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      type: object
    platform.Envelope:
      properties:
        data: {}
        error:
          $ref: '#/components/schemas/platform.APIError'
        request_id:
          type: string
      type: object
...
    service.EventDetail:
      properties:
        cover_image_key:
          type: string
        cover_image_url:
          type: string
        created_at:
          type: string
        description:
          type: string
        ends_at:
          type: string
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        starts_at:
          type: string
        status:
          type: string
        timezone:
          type: string
        venue_address:
          type: string
        venue_name:
          type: string
      type: object
...
    service.TicketTypeDetail:
      properties:
        capacity:
          type: integer
        created_at:
          type: string
        currency:
          type: string
        description:
          type: string
        event_id:
          type: string
        id:
          type: string
        name:
          type: string
        price_cents:
          type: integer
        sold_count:
          type: integer
        sort_order:
          type: integer
        updated_at:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Staff session token as Bearer <session_id>. The Staff BFF also
        accepts an httpOnly cookie.
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: |-
    Ticket POS backend API.
    JSON endpoints use the standard envelope with data, error, and request_id.
    /health is the only endpoint that returns a bare JSON object.
  title: Ticket POS API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/staff/events/{id}:
    delete:
      description: Deletes a draft catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeMessage'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - staff
    get:
      description: Returns a catalog event by ID.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - staff
    patch:
      description: Updates catalog event fields.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.updateEventBody'
                description: Event fields
                summary: body
        description: Event fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeEventDetail'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - staff
...
  /api/v1/staff/events/{id}/ticket-types:
    get:
      description: Lists ticket types for a catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeTicketTypeList'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: List ticket types
      tags:
      - staff
    post:
      description: Creates a ticket type on a catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.createTicketTypeBody'
                description: Ticket type details
                summary: body
        description: Ticket type details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeTicketTypeDetail'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Create ticket type
      tags:
      - staff
  /api/v1/staff/events/{id}/ticket-types/{ticketTypeId}:
    delete:
      description: Deletes a ticket type when the parent event is a draft.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Ticket type ID
        in: path
        name: ticketTypeId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeMessage'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Conflict
      security:
      - BearerAuth: []
      summary: Delete ticket type
      tags:
      - staff
    patch:
      description: Updates a ticket type on a catalog event.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Ticket type ID
        in: path
        name: ticketTypeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.updateTicketTypeBody'
                description: Ticket type fields
                summary: body
        description: Ticket type fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/openapi.EnvelopeTicketTypeDetail'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/platform.Envelope'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update ticket type
      tags:
      - staff
...